+++ coordinator/migrations/2023-10-16-113322_add_stable_flag/down.sql
@@ -1,0 +1,6 @@
+-- This file should undo anything in `up.sql`
+ALTER TABLE orders
+    DROP COLUMN "stable";
+
+ALTER TABLE positions
+    DROP COLUMN "stable";