+++ crates/tests-e2e/src/setup.rs
@@ -109,5 +109,6 @@ pub fn dummy_order() -> NewOrder {
         direction: api::Direction::Long,
         quantity: 1.0,
         order_type: Box::new(OrderType::Market),
+        stable: false,
     }
 }